Move Chart into a separate, shared, package.
