+++ jira_analysis/cycle_time/chart/cycle_time/scatter.py
@@ -4,7 +4,7 @@ from bokeh.models.sources import DataSource
 from collections import Counter
 from typing import List, Type
 
-from jira_analysis.cycle_time.chart.base import IChart, Plot
+from jira_analysis.chart.base import IChart, Plot
 from jira_analysis.cycle_time.cycle_time import CycleTime
 from .utils import sort_cycle_times, unsplit
