+++ test/cycle_time/chart/test_ticket_control.py
@@ -4,7 +4,7 @@ from datetime import datetime
 from unittest import mock
 
 from jira_analysis.cycle_time.issue import Issue
-from jira_analysis.cycle_time.chart.base import Chart
+from jira_analysis.chart.base import Chart
 from jira_analysis.cycle_time.chart.ticket_control import generate_control_chart
